requirements.txt
ctransformers==0.2.5
sentence-transformers==2.4.0
pinecone-client
pinecone-client[grpc]
langchain
flask
pypdf
python-dotenv
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
-e .